swaybar: Implement pointer gesture swipe support
After previous commits, swaybar bindsym statements would already accept
the new pseudo button names for swipe gestures. Also, sway itself would
not handle swipes when executed while the pointer was above a bar.

Add the necessary handling for three- and four-finger swipe pointer
gestures in swaybar. Add the pointer gestures protocol in client and
server variants to the meson build file to make the necessary
definitions available. Bind to the global interface on startup,
instantiate a swipe gesture and add a listener on seat setup as well as
destroy it on seat shutdown.

Extend the sway-bar manual page as necessary.

Test-plan:
- add workspace switching to config like so:

bar bar-0 {
        swaybar_command swaybar

        bindsym SWIPE_3_LEFT exec yad --text foo
        bindsym SWIPE_4_RIGHT exec yad --text bar
        bindsym --release SWIPE_3_LEFT workspace prev_on_output
        bindsym --release SWIPE_4_RIGHT workspace next_on_output
}

- start sway and open two workspaces
- position pointer above a bar surface
- switch back and forth using horizontal three- and four-finger swipes,
  observing that different finger counts are necessary per direction but the
  respective yad popup appears on any swipe of the respective finger count, no
  matter what direction it's initially going.
- move pointer away from bar surface
- observe that switching by swipe no longer works

Signed-off-by: Michael Weiser <michael.weiser@gmx.de>