criteria: Allow re-execution of criteria
Currently, criteria are executed only once. This change keeps the
principal behaviour of not re-executing criteria while their matching
state has not changed. But as soon as a criterion no longer matches, it
will be removed from the list of already executed criteria. This allows
its action to be re-executed once the criterion once again matches the
view.

Since function criteria_for_view() returns a list of all currently
matching criteria, this can be trivially implemented and is even
slightly more efficient than before: Instead of checking each matching
criterion against the list of already executed criteria and potentially
executing and adding it, we just make the list returned by
criteria_for_view() the new list of executed criteria. Then we walk
through it and execute every criterion not already in the old list of
exectued criteria. Adding the criterion to the list becomes redundant.
The old list is freed at the end.

With this approach we indirectly forget about any criteria we may have
executed in the past but who no longer match the view right now.

As an additional simplification, function view_has_executed_criteria()
is replaced by list_find() since it does the same thing.

Since execution behaviour was not documented explicitly before, we add a
note to sway(5) on how it is behaving now.

Signed-off-by: Michael Weiser <michael.weiser@gmx.de>